Extracted delete conversation
